Integrate server-poll to employee database.
I have integrated the server-poll functionality to the
employee_database. I plan to add network functionality to the
employee_database. I have extended the network protocol to work for
sending hello messages in order to initiate communication between server
and client, adding an employee to the database, list employees in the
database, and deleting an employee from the database.